+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Micro Manager changelog
 
+## v0.11.2
+
+- Fixed use of provided HDF5 output file name when snapshots are computed in parallel. [#315](https://github.com/precice/micro-manager/pull/315)
+- Added various fixes for interpolation, load balancing and adaptivity. Load balancing now balances every N implicit iterations. [#312](https://github.com/precice/micro-manager/pull/312)
+
 ## v0.11.0
 
 - Refactored LoadBalancing into an interface for more clarity [#300](https://github.com/precice/micro-manager/pull/300)